initial multiple users for a db

+++ modules/fbs-core/api/src/main/kotlin/de/thm/ii/fbs/controller/v2/PlaygroundController.kt
@@ -7,10 +7,12 @@ import de.thm.ii.fbs.model.v2.playground.*
 import de.thm.ii.fbs.model.v2.playground.api.SqlPlaygroundDatabaseCreation
 import de.thm.ii.fbs.model.v2.playground.api.SqlPlaygroundQueryCreation
 import de.thm.ii.fbs.model.v2.playground.api.SqlPlaygroundResult
+import de.thm.ii.fbs.model.v2.playground.api.SqlPlaygroundUsersCreation
 import de.thm.ii.fbs.services.v2.checker.SqlPlaygroundCheckerService
 import de.thm.ii.fbs.services.v2.persistence.*
 import de.thm.ii.fbs.utils.v2.annotations.CurrentToken
 import de.thm.ii.fbs.utils.v2.exceptions.NotFoundException
+import de.thm.ii.fbs.utils.v2.exceptions.UnauthorizedException
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.*
 
@@ -22,6 +24,7 @@ class PlaygroundController(
         private val entityRepository: SqlPlaygroundEntityRepository,
         private val queryRepository: SqlPlaygroundQueryRepository,
         private val sqlPlaygroundCheckerService: SqlPlaygroundCheckerService,
+        private val sqlPlaygroundUsersRepository: SqlPlaygroundUsersRepository,
 ) {
     @GetMapping
     @ResponseBody
@@ -44,23 +47,54 @@ class PlaygroundController(
     @DeleteMapping("/{dbId}")
     @ResponseBody
     fun delete(@CurrentToken currentToken: LegacyToken, @PathVariable("dbId") dbId: Int): SqlPlaygroundDatabase {
-        val db = databaseRepository.findByOwner_IdAndIdAndDeleted(currentToken.id, dbId, false) ?: throw NotFoundException()
-        sqlPlaygroundCheckerService.deleteDatabase(db, currentToken.id, currentToken.username)
-        db.active = false
-        db.deleted = true
-        databaseRepository.save(db)
-        return db
+        val db = databaseRepository.findByOwner_IdAndIdAndDeleted(currentToken.id, dbId, false)
+                ?: throw NotFoundException()
+        if (currentToken.id == db.owner.id) {
+            sqlPlaygroundCheckerService.deleteDatabase(db, currentToken.id, currentToken.username)
+            db.active = false
+            db.deleted = true
+            databaseRepository.save(db)
+            return db
+        }
+        throw UnauthorizedException()
+    }
+
+    @PostMapping("/dbusers")
+    @ResponseBody
+    fun addUserToDB(@CurrentToken currentToken: LegacyToken, @RequestBody userCreation: SqlPlaygroundUsersCreation): SqlPlaygroundUsers {
+        val db = databaseRepository.findById(userCreation.dbId).get()
+        if (currentToken.id == db.owner.id) {
+            return sqlPlaygroundUsersRepository.save(SqlPlaygroundUsers(user = userRepository.findById(userCreation.userId).get(), db = db));
+        }
+        throw UnauthorizedException()
+    }
+
+    @DeleteMapping("/dbusers/{id}")
+    @ResponseBody
+    fun removeUserFromDB(@CurrentToken currentToken: LegacyToken, @PathVariable id: Int, @RequestBody userCreation: SqlPlaygroundUsersCreation): Unit {
+        val db = databaseRepository.findById(userCreation.dbId).get()
+        if (currentToken.id == db.owner.id) {
+            return sqlPlaygroundUsersRepository.deleteById(id)
+        }
+        throw UnauthorizedException()
+    }
+
+    @GetMapping("/dbusers/all")
+    @ResponseBody
+    fun getAllDBsOfUser(@CurrentToken currentToken: LegacyToken): List<SqlPlaygroundDatabase> {
+        return sqlPlaygroundUsersRepository.findAllSqlPlaygroundDatabasesByUserId(currentToken.id)
     }
 
     @GetMapping("/{dbId}")
     @ResponseBody
     fun get(@CurrentToken currentToken: LegacyToken, @PathVariable("dbId") dbId: Int): SqlPlaygroundDatabase =
-        databaseRepository.findByOwner_IdAndIdAndDeleted(currentToken.id, dbId, false) ?: throw NotFoundException()
+            databaseRepository.findByOwner_IdAndIdAndDeleted(currentToken.id, dbId, false) ?: throw NotFoundException()
 
     @PostMapping("/{dbId}/activate")
     @ResponseBody
     fun activate(@CurrentToken currentToken: LegacyToken, @PathVariable("dbId") dbId: Int): SqlPlaygroundDatabase {
-        val db = databaseRepository.findByOwner_IdAndIdAndDeleted(currentToken.id, dbId, false) ?: throw NotFoundException()
+        val db = databaseRepository.findByOwner_IdAndIdAndDeleted(currentToken.id, dbId, false)
+                ?: throw NotFoundException()
         db.active = true
         val currentActiveDb = databaseRepository.findByOwner_IdAndActiveAndDeleted(currentToken.id, true, false)
         if (currentActiveDb !== null) {
@@ -79,8 +113,10 @@ class PlaygroundController(
     @ResponseBody
     @ResponseStatus(HttpStatus.ACCEPTED)
     fun execute(@CurrentToken currentToken: LegacyToken, @PathVariable("dbId") dbId: Int, @RequestBody sqlQuery: SqlPlaygroundQueryCreation): SqlPlaygroundQuery {
-        val db = databaseRepository.findByOwner_IdAndIdAndDeleted(currentToken.id, dbId, false) ?: throw NotFoundException()
-        val query = queryRepository.save(SqlPlaygroundQuery(sqlQuery.statement, db))
+
+        val db = sqlPlaygroundUsersRepository.findSqlPlaygroundDatabasesByMemberIdAndDBId(currentToken.id, dbId)
+                ?: throw NotFoundException()
+        val query = queryRepository.save(SqlPlaygroundQuery(sqlQuery.statement, db, currentToken.id))
         sqlPlaygroundCheckerService.submit(query)
         return query
     }
@@ -88,42 +124,44 @@ class PlaygroundController(
     @GetMapping("/{dbId}/results")
     @ResponseBody
     fun getResults(@CurrentToken currentToken: LegacyToken, @PathVariable("dbId") dbId: Int): List<SqlPlaygroundResult> =
-            queryRepository.findByRunIn_Owner_IdAndRunIn_id(currentToken.id, dbId).mapNotNull { it.result }
+            queryRepository.findAllByCreatorIdAndDatabaseId(currentToken.id, dbId).mapNotNull { it.result }
 
     @GetMapping("/{dbId}/results/{qId}")
     @ResponseBody
     fun getResult(@CurrentToken currentToken: LegacyToken, @PathVariable("dbId") dbId: Int, @PathVariable("qId") qId: Int): SqlPlaygroundResult =
-            queryRepository.findByRunIn_Owner_IdAndRunIn_idAndId(currentToken.id, dbId, qId)?.result  ?: throw NotFoundException()
+            queryRepository.findByRunIn_Creator_IdAndRunIn_idAndId(currentToken.id, dbId, qId)?.result
+                    ?: throw NotFoundException()
 
     @GetMapping("/{dbId}/tables")
     @ResponseBody
     fun getTables(@CurrentToken currentToken: LegacyToken, @PathVariable("dbId") dbId: Int): ArrayNode =
-        getEntity(currentToken.id, dbId, "tables")
+            getEntity(currentToken.id, dbId, "tables")
 
     @GetMapping("/{dbId}/constraints")
     @ResponseBody
     fun getConstrains(@CurrentToken currentToken: LegacyToken, @PathVariable("dbId") dbId: Int): ArrayNode =
-        getEntity(currentToken.id, dbId, "constraints")
+            getEntity(currentToken.id, dbId, "constraints")
 
     @GetMapping("/{dbId}/views")
     @ResponseBody
     fun getViews(@CurrentToken currentToken: LegacyToken, @PathVariable("dbId") dbId: Int): ArrayNode =
-        getEntity(currentToken.id, dbId, "views")
+            getEntity(currentToken.id, dbId, "views")
 
     @GetMapping("/{dbId}/routines")
     @ResponseBody
     fun getRoutines(@CurrentToken currentToken: LegacyToken, @PathVariable("dbId") dbId: Int): ArrayNode =
-        getEntity(currentToken.id, dbId, "routines")
+            getEntity(currentToken.id, dbId, "routines")
 
     @GetMapping("/{dbId}/triggers")
     @ResponseBody
     fun getTriggers(@CurrentToken currentToken: LegacyToken, @PathVariable("dbId") dbId: Int): ArrayNode =
-        getEntity(currentToken.id, dbId, "triggers")
+            getEntity(currentToken.id, dbId, "triggers")
 
     private fun getEntity(userId: Int, databaseId: Int, type: String) =
-        entityRepository.findByDatabase_Owner_IdAndDatabase_idAndDatabase_DeletedAndType(userId, databaseId, false, type)?.data ?: throw NotFoundException()
+            entityRepository.findByDatabase_Owner_IdAndDatabase_idAndDatabase_DeletedAndType(userId, databaseId, false, type)?.data
+                    ?: throw NotFoundException()
 
-    private fun createAllEntities(database: SqlPlaygroundDatabase) = listOf("tables", "constraints", "views", "routines", "triggers").forEach {type ->
+    private fun createAllEntities(database: SqlPlaygroundDatabase) = listOf("tables", "constraints", "views", "routines", "triggers").forEach { type ->
         entityRepository.save(SqlPlaygroundEntity(database, type, ArrayNode(JsonNodeFactory(false))))
     }
 }
